+++ docs/Changelog.md
@@ -22,7 +22,7 @@
   - `Parameters.POLY_SYMBOL = true;`
 - Map fractional zoom [#487](https://github.com/mapsforge/vtm/issues/487)
 - Render theme fallback internal resources [#477](https://github.com/mapsforge/vtm/issues/477)
-- Fix ItemizedLayer synchronization [#507](https://github.com/mapsforge/vtm/issues/507)
+- Fix layers synchronization [#507](https://github.com/mapsforge/vtm/issues/507)
 - Fix Mapsforge clipping on small zooms [#264](https://github.com/mapsforge/vtm/issues/264)
 - Fix PathLayer (vtm) reverse segments [#220](https://github.com/mapsforge/vtm/issues/220)
 - Fix FadeStep alpha interpolation [#486](https://github.com/mapsforge/vtm/issues/486)